Rework to use Rails-native sessions
The Cognito example used in initial setup stored every Cognito session via ActiveRecord.
This was fine, b/c we'd have to lookup user sessions to see if they had expired.
In testing/development, this doesn't matter, and just creates a lot of cruft.
In production, we'll be applying the Cognito user pool directly to the application
load balancer, so they won't ever reach our app if they don't have an active, valid session.